requirements.txt
pandas==2.3.3
openpyxl==3.1.5
djangorestframework==3.16.1
django-filter==25.2
plotly==6.5.2
